+++ disdrodb/utils/time.py
@@ -670,133 +670,3 @@ def regularize_timesteps(ds, sample_interval, robust=False, add_quality_flag=Tru
         ds = ds.isel(time=idx_valid_timesteps)
     # Return dataset
     return ds
-
-
-####---------------------------------------------------------------------------------
-#### Time blocks
-
-
-def check_freq(freq: str) -> None:
-    """Check validity of freq argument."""
-    valid_freq = ["none", "year", "season", "quarter", "month", "day", "hour"]
-    if not isinstance(freq, str):
-        raise TypeError("'freq' must be a string.")
-    if freq not in valid_freq:
-        raise ValueError(
-            f"'freq' '{freq}' is not possible. Must be one of: {valid_freq}.",
-        )
-    return freq
-
-
-def generate_time_blocks(start_time: np.datetime64, end_time: np.datetime64, freq: str) -> np.ndarray:  # noqa: PLR0911
-    """Generate time blocks between `start_time` and `end_time` for a given frequency.
-
-    Parameters
-    ----------
-    start_time : numpy.datetime64
-        Inclusive start of the overall time range.
-    end_time : numpy.datetime64
-        Inclusive end of the overall time range.
-    freq : str
-        Frequency specifier. Accepted values are:
-        - 'none'    : return a single block [start_time, end_time]
-        - 'day'     : split into daily blocks
-        - 'month'   : split into calendar months
-        - 'quarter' : split into calendar quarters
-        - 'year'    : split into calendar years
-        - 'season'  : split into meteorological seasons (MAM, JJA, SON, DJF)
-
-    Returns
-    -------
-    numpy.ndarray
-        Array of shape (n, 2) with dtype datetime64[s], where each row is [block_start, block_end].
-
-    """
-    freq = check_freq(freq)
-    if freq == "none":
-        return np.array([[start_time, end_time]], dtype="datetime64[s]")
-
-    if freq == "hour":
-        periods = pd.period_range(start=start_time, end=end_time, freq="h")
-        blocks = np.array(
-            [
-                [
-                    period.start_time.to_datetime64().astype("datetime64[s]"),
-                    period.end_time.to_datetime64().astype("datetime64[s]"),
-                ]
-                for period in periods
-            ],
-            dtype="datetime64[s]",
-        )
-        return blocks
-
-    if freq == "day":
-        periods = pd.period_range(start=start_time, end=end_time, freq="d")
-        blocks = np.array(
-            [
-                [
-                    period.start_time.to_datetime64().astype("datetime64[s]"),
-                    period.end_time.to_datetime64().astype("datetime64[s]"),
-                ]
-                for period in periods
-            ],
-            dtype="datetime64[s]",
-        )
-        return blocks
-
-    if freq == "month":
-        periods = pd.period_range(start=start_time, end=end_time, freq="M")
-        blocks = np.array(
-            [
-                [
-                    period.start_time.to_datetime64().astype("datetime64[s]"),
-                    period.end_time.to_datetime64().astype("datetime64[s]"),
-                ]
-                for period in periods
-            ],
-            dtype="datetime64[s]",
-        )
-        return blocks
-
-    if freq == "year":
-        periods = pd.period_range(start=start_time, end=end_time, freq="Y")
-        blocks = np.array(
-            [
-                [
-                    period.start_time.to_datetime64().astype("datetime64[s]"),
-                    period.end_time.to_datetime64().astype("datetime64[s]"),
-                ]
-                for period in periods
-            ],
-            dtype="datetime64[s]",
-        )
-        return blocks
-
-    if freq == "quarter":
-        periods = pd.period_range(start=start_time, end=end_time, freq="Q")
-        blocks = np.array(
-            [
-                [
-                    period.start_time.to_datetime64().astype("datetime64[s]"),
-                    period.end_time.floor("s").to_datetime64().astype("datetime64[s]"),
-                ]
-                for period in periods
-            ],
-            dtype="datetime64[s]",
-        )
-        return blocks
-
-    # if freq == "season":
-    # Fiscal quarter frequency ending in Feb → seasons DJF, MAM, JJA, SON
-    periods = pd.period_range(start=start_time, end=end_time, freq="Q-FEB")
-    blocks = np.array(
-        [
-            [
-                period.start_time.to_datetime64().astype("datetime64[s]"),
-                period.end_time.to_datetime64().astype("datetime64[s]"),
-            ]
-            for period in periods
-        ],
-        dtype="datetime64[s]",
-    )
-    return blocks
